refactor: Tidy `PropertiesPermutationHelper`
- Explain what the class is actually for.
- Make `ApplyPermutation` a pure function.
- Move implementation into matching .cpp file for its .hpp
- Correct the naming convention of public methods.
