+++ src/storage/v2/indices/label_property_index.cpp
@@ -11,7 +11,85 @@
 
 #include "label_property_index.hpp"
 
+namespace r = ranges;
+namespace rv = r::views;
+
 namespace memgraph::storage {
+namespace {
+
+auto build_permutation_cycles(std::span<std::size_t const> permutation_index)
+    -> PropertiesPermutationHelper::permutation_cycles {
+  auto const n = permutation_index.size();
+
+  auto visited = std::vector(n, false);
+  auto cycle = std::vector<std::size_t>{};
+  auto cycles = PropertiesPermutationHelper::permutation_cycles{};
+
+  for (auto i = std::size_t{}; i != n; ++i) {
+    if (visited[i]) [[unlikely]] {
+      // already part of a cycle
+      continue;
+    }
+
+    // build a cycle
+    cycle.clear();
+    auto current = i;
+    do {
+      visited[current] = true;
+      cycle.push_back(current);
+      current = permutation_index[current];
+    } while (current != i);
+
+    if (cycle.size() == 1) {
+      // Ignore self-mapped elements
+      continue;
+    }
+    cycles.emplace_back(std::move(cycle));
+  }
+  return cycles;
+}
+}  // end namespace
+
+PropertiesPermutationHelper::PropertiesPermutationHelper(std::span<PropertyId const> properties)
+    : sorted_properties_(properties.begin(), properties.end()) {
+  auto inverse_permutation = rv::iota(size_t{}, properties.size()) | r::to_vector;
+  r::sort(rv::zip(inverse_permutation, sorted_properties_), std::less{},
+          [](auto const &value) -> decltype(auto) { return (std::get<1>(value)); });
+  position_lookup_ = std::move(inverse_permutation);
+  cycles_ = build_permutation_cycles(position_lookup_);
+}
+
+auto PropertiesPermutationHelper::Extract(PropertyStore const &properties) const -> std::vector<PropertyValue> {
+  return properties.ExtractPropertyValuesMissingAsNull(sorted_properties_);
+}
+
+auto PropertiesPermutationHelper::ApplyPermutation(std::vector<PropertyValue> values) const
+    -> IndexOrderedPropertyValues {
+  for (const auto &cycle : cycles_) {
+    auto tmp = std::move(values[cycle.front()]);
+    for (auto pos : std::span{cycle}.subspan<1>()) {
+      tmp = std::exchange(values[pos], tmp);
+    }
+    values[cycle.front()] = std::move(tmp);
+  }
+
+  return {std::move(values)};
+}
+
+auto PropertiesPermutationHelper::MatchesValue(PropertyId property_id, PropertyValue const &value,
+                                               IndexOrderedPropertyValues const &values) const
+    -> std::optional<std::pair<std::ptrdiff_t, bool>> {
+  auto it = std::ranges::find(sorted_properties_, property_id);
+  if (it == sorted_properties_.end()) return std::nullopt;
+
+  auto pos = std::distance(sorted_properties_.begin(), it);
+  return std::pair{pos, values.values_[position_lookup_[pos]] == value};
+}
+
+auto PropertiesPermutationHelper::MatchesValues(PropertyStore const &properties,
+                                                IndexOrderedPropertyValues const &values) const -> std::vector<bool> {
+  return properties.ArePropertiesEqual(sorted_properties_, values.values_, position_lookup_);
+}
 
 size_t PropertyValueRange::hash() const noexcept {
   auto const lower{lower_ ? lower_->value() : PropertyValue{}};
